// packages/apps/bingo/src/types.ts
export type TileItem = { text: string; active: boolean };

export type ThemeConfig = {
  appBgImg: string;
  logoImg: string;
  logoImgWidth: number;
  appBackgroundColor: string;
  tileBg: string;
  tileBgActive: string;
  tileBorder: string;
  tileBorderActive: string;
  tileText: string;
  tileTextActive: string;
};